Refactored submodule to bring main functionalities to the top Toeplitz level.
